Separate geocoding stuff from modem to Geocoding trait
